Usage description for next.js
Add a usage section to the readme, targeted for users of Nextjs+MDX. Explains the dependency on `react-slug`

My apologies if you find this inappropriate for this library, because I know it is meant for use with unified. I just think it will save folks some time because this plugin is very popular in next apps, and I was confused for days when it did not work without `remark-slug`!

As an alternative, maybe add an error message in the case where remark-slug is not installed? Currently this plugin is silently broken without it.